backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   -r requirements.in
    #   keras
    #   tensorboard
    #   tensorflow
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
astunparse==1.6.3
    # via
    #   -r requirements.in
    #   tensorflow
attrs==25.1.0
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements.in
    #   requests
django==5.0.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.6.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
flatbuffers==25.1.24
    # via
    #   -r requirements.in
    #   tensorflow
gast==0.6.0
    # via
    #   -r requirements.in
    #   tensorflow
google-pasta==0.2.0
    # via
    #   -r requirements.in
    #   tensorflow
grpcio==1.70.0
    # via
    #   -r requirements.in
    #   tensorboard
    #   tensorflow
gunicorn==23.0.0
    # via -r requirements.in
h5py==3.12.1
    # via
    #   -r requirements.in
    #   keras
    #   tensorflow
idna==3.10
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema==4.23.0
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements.in
    #   jsonschema
keras==3.8.0
    # via
    #   -r requirements.in
    #   tensorflow
libclang==18.1.1
    # via
    #   -r requirements.in
    #   tensorflow
markdown==3.7
    # via
    #   -r requirements.in
    #   tensorboard
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   werkzeug
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
ml-dtypes==0.3.2
    # via
    #   -r requirements.in
    #   keras
    #   tensorflow
namex==0.0.8
    # via
    #   -r requirements.in
    #   keras
numpy==1.26.4
    # via
    #   -r requirements.in
    #   h5py
    #   keras
    #   ml-dtypes
    #   tensorboard
    #   tensorflow
opt-einsum==3.4.0
    # via
    #   -r requirements.in
    #   tensorflow
optree==0.14.0
    # via
    #   -r requirements.in
    #   keras
packaging==24.2
    # via
    #   -r requirements.in
    #   gunicorn
    #   keras
    #   tensorflow
pillow==11.1.0
    # via -r requirements.in
protobuf==4.25.6
    # via
    #   -r requirements.in
    #   tensorboard
    #   tensorflow
psycopg2-binary==2.9.10
    # via -r requirements.in
pygments==2.19.1
    # via
    #   -r requirements.in
    #   rich
pyjwt==2.10.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   drf-spectacular
referencing==0.36.2
    # via
    #   -r requirements.in
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   tensorflow
rich==13.9.4
    # via
    #   -r requirements.in
    #   keras
rpds-py==0.22.3
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   -r requirements.in
    #   astunparse
    #   google-pasta
    #   tensorboard
    #   tensorflow
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tensorboard==2.16.2
    # via
    #   -r requirements.in
    #   tensorflow
tensorboard-data-server==0.7.2
    # via
    #   -r requirements.in
    #   tensorboard
tensorflow==2.16.2
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.37.1
    # via
    #   -r requirements.in
    #   tensorflow
termcolor==2.5.0
    # via
    #   -r requirements.in
    #   tensorflow
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   asgiref
    #   optree
    #   referencing
    #   rich
    #   tensorflow
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   drf-spectacular
urllib3==2.3.0
    # via
    #   -r requirements.in
    #   requests
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   tensorboard
wheel==0.45.1
    # via astunparse
whitenoise==6.8.2
    # via -r requirements.in
wrapt==1.17.2
    # via
    #   -r requirements.in
    #   tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
